move next data path to utils
